update utils/vcard/src/vcard_encoder.cpp. 使用size_t代替int_32的强转换，去除编译警告

+++ utils/vcard/src/vcard_encoder.cpp
@@ -100,9 +100,9 @@ std::shared_ptr<DataShare::DataShareResultSet> VCardEncoder::GetRawContactResult
     columns.push_back(RawContact::ID);
     DataShare::DataSharePredicates predicates;
     predicates.BeginWrap();
-    for (int i = 0; i < (int32_t)contactIdList.size(); i++) {
+    for (size_t i = 0; i < contactIdList.size(); i++) {
         predicates.EqualTo(RawContact::CONTACT_ID, contactIdList[i]);
-        if (i != (int32_t)contactIdList.size() - 1) {
+        if (i != contactIdList.size() - 1) {
             predicates.Or();
         }
     }
